+++ internal/commands/push_test.go
@@ -101,14 +101,14 @@ func TestGenerateNewID(t *testing.T) {
 		return time.Date(2023, time.November, 10, 12, 32, 43, 0, time.UTC)
 	}
 
-	id := generateNewId(now, &model.ThingModel{
+	id, _ := generateNewId(now, &model.ThingModel{
 		Manufacturer: model.SchemaManufacturer{Name: "omnicorp"},
 		Mpn:          "senseall",
 		Author:       model.SchemaAuthor{Name: "author"},
 		Version:      model.Version{Model: "v3.2.1"},
 	}, []byte("{\n\"title\":\"test\"\n}"), "opt/dir")
 
-	assert.Equal(t, "author/omnicorp/senseall/opt/dir/v3.2.1-20231110123243-863e9f0f950a.tm.json", id.String())
+	assert.Equal(t, "author/omnicorp/senseall/opt/dir/v3.2.1-20231110123243-7ae21a619c71.tm.json", id.String())
 }
 
 func TestPrepareToImport(t *testing.T) {
@@ -124,7 +124,7 @@ func TestPrepareToImport(t *testing.T) {
 		}, []byte("{\r\n\"title\":\"test\"\r\n}"), "opt/dir")
 		assert.NoError(t, err)
 		assert.False(t, bytes.Contains(b, []byte{'\r'})) // make sure line endings were normalized
-		assert.True(t, bytes.Contains(b, []byte("author/omnicorp/senseall/opt/dir/v3.2.1-20231110123243-863e9f0f950a.tm.json")))
+		assert.True(t, bytes.Contains(b, []byte("author/omnicorp/senseall/opt/dir/v3.2.1-20231110123243-7ae21a619c71.tm.json")))
 	})
 	t.Run("foreign string id in original", func(t *testing.T) {
 		b, _, err := prepareToImport(now, &model.ThingModel{
@@ -135,7 +135,7 @@ func TestPrepareToImport(t *testing.T) {
 		}, []byte("{\r\n\"title\":\"test\"\r\n,\"id\":\"foreign-id\"}"), "opt/dir")
 		assert.NoError(t, err)
 		assert.True(t, bytes.Contains(b, []byte("\"href\":\"foreign-id\"")))
-		assert.True(t, bytes.Contains(b, []byte("author/omnicorp/senseall/opt/dir/v3.2.1-20231110123243-e8ed3d810b9a.tm.json")))
+		assert.True(t, bytes.Contains(b, []byte("author/omnicorp/senseall/opt/dir/v3.2.1-20231110123243-e7dac5728be6.tm.json")))
 	})
 	t.Run("our string id in original/correct hash", func(t *testing.T) {
 		b, _, err := prepareToImport(now, &model.ThingModel{
@@ -143,9 +143,9 @@ func TestPrepareToImport(t *testing.T) {
 			Mpn:          "senseall",
 			Author:       model.SchemaAuthor{Name: "author"},
 			Version:      model.Version{Model: "v3.2.1"},
-		}, []byte("{\r\n\"title\":\"test\"\r\n,\"id\":\"author/omnicorp/senseall/opt/dir/v3.2.1-20231010123243-60afcb846e38.tm.json\"}"), "opt/dir")
+		}, []byte("{\r\n\"title\":\"test\"\r\n,\"id\":\"author/omnicorp/senseall/opt/dir/v3.2.1-20231010123243-7ae21a619c71.tm.json\"}"), "opt/dir")
 		assert.NoError(t, err)
-		assert.True(t, bytes.Contains(b, []byte("author/omnicorp/senseall/opt/dir/v3.2.1-20231010123243-60afcb846e38.tm.json")))
+		assert.True(t, bytes.Contains(b, []byte("author/omnicorp/senseall/opt/dir/v3.2.1-20231010123243-7ae21a619c71.tm.json")))
 	})
 	t.Run("our string id in original/incorrect hash", func(t *testing.T) {
 		b, _, err := prepareToImport(now, &model.ThingModel{
@@ -155,7 +155,7 @@ func TestPrepareToImport(t *testing.T) {
 			Version:      model.Version{Model: "v3.2.1"},
 		}, []byte("{\r\n\"title\":\"test\"\r\n,\"id\":\"author/omnicorp/senseall/opt/dir/v3.2.1-20231010123243-863e9f0f950a.tm.json\"}"), "opt/dir")
 		assert.NoError(t, err)
-		assert.True(t, bytes.Contains(b, []byte("author/omnicorp/senseall/opt/dir/v3.2.1-20231110123243-60afcb846e38.tm.json")))
+		assert.True(t, bytes.Contains(b, []byte("author/omnicorp/senseall/opt/dir/v3.2.1-20231110123243-7ae21a619c71.tm.json")))
 	})
 
 }