+++ internal/commands/versions.go
@@ -16,26 +16,26 @@ func NewVersionsCommand(manager remotes.RemoteManager) *VersionsCommand {
 		remoteMgr: manager,
 	}
 }
-func (c *VersionsCommand) ListVersions(spec remotes.RepoSpec, name string) (model.FoundEntry, error) {
+func (c *VersionsCommand) ListVersions(spec remotes.RepoSpec, name string) ([]model.FoundVersion, error) {
 	rs, err := remotes.GetSpecdOrAll(c.remoteMgr, spec)
 	if err != nil {
-		return model.FoundEntry{}, err
+		return nil, err
 	}
-	res := model.FoundEntry{}
+	var res []model.FoundVersion
 	found := false
 	for _, remote := range rs {
-		toc, err := remote.Versions(name)
+		vers, err := remote.Versions(name)
 		if err != nil && errors.Is(err, remotes.ErrEntryNotFound) {
 			continue
 		}
 		if err != nil {
-			return model.FoundEntry{}, err
+			return nil, err
 		}
 		found = true
-		res = res.Merge(toc)
+		res = model.MergeFoundVersions(res, vers)
 	}
 	if !found {
-		return model.FoundEntry{}, remotes.ErrEntryNotFound
+		return nil, remotes.ErrEntryNotFound
 	}
 	return res, nil