+++ internal/remotes/mock_Remote.go
@@ -138,22 +138,24 @@ func (_m *MockRemote) UpdateToc(updatedFiles ...string) error {
 }
 
 // Versions provides a mock function with given fields: name
-func (_m *MockRemote) Versions(name string) (model.FoundEntry, error) {
+func (_m *MockRemote) Versions(name string) ([]model.FoundVersion, error) {
 	ret := _m.Called(name)
 
 	if len(ret) == 0 {
 		panic("no return value specified for Versions")
 	}
 
-	var r0 model.FoundEntry
+	var r0 []model.FoundVersion
 	var r1 error
-	if rf, ok := ret.Get(0).(func(string) (model.FoundEntry, error)); ok {
+	if rf, ok := ret.Get(0).(func(string) ([]model.FoundVersion, error)); ok {
 		return rf(name)
 	}
-	if rf, ok := ret.Get(0).(func(string) model.FoundEntry); ok {
+	if rf, ok := ret.Get(0).(func(string) []model.FoundVersion); ok {
 		r0 = rf(name)
 	} else {
-		r0 = ret.Get(0).(model.FoundEntry)
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).([]model.FoundVersion)
+		}
 	}
 
 	if rf, ok := ret.Get(1).(func(string) error); ok {