+++ CHANGELOG.md
@@ -6,12 +6,16 @@
 
 - Implemented file locking to avoid data races (guaranteed for Linux/MacOSX)
 - Implemented ```pull``` command to fetch and save multiple thing models at once 
+- Implemented a 'tmc' remote type, which uses our own REST API as the underlying TM storage
+- Implemented ```digest``` command to calculate the version digest of a TM file
 
 ### Changed
 
 - ```create-toc```: renamed to ```update-toc``` and allow for partial updates
 - ```list```: allows now listing by name pattern
 - ```serve```: separate configuration of the remote(s) to be served from the target remote for push  
+- (BREAKING!) ```push```: file hash calculation has been made more reliable and idempotent. Consequently, some files if pushed to TMC, may receive a new version hash, despite no change in contents
+
 
 ### Fixed